move massive view-dependent API data filtering function, up from YearLimitsSelect component, to utils directory
