+++ src/utils/index.js
@@ -1,0 +1,30 @@
+import { rawApiDataToPlotlyReadyInfo } from './rawApiDataToPlotlyReadyInfo';
+
+/* 
+  useInterval here is from https://overreacted.io/making-setinterval-declarative-with-react-hooks/ 
+    -- Mack
+*/
+
+import { useEffect, useRef } from 'react';
+
+function useInterval(callback, delay) {
+  const savedCallback = useRef();
+
+  // Remember the latest callback.
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  // Set up the interval.
+  useEffect(() => {
+    function tick() {
+      savedCallback.current();
+    }
+    if (delay !== null) {
+      let id = setInterval(tick, delay);
+      return () => clearInterval(id);
+    }
+  }, [delay]);
+}
+
+export { useInterval, rawApiDataToPlotlyReadyInfo };
